backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
anthropic==0.7.8
sentence-transformers==2.7.0
chromadb==0.4.18
numpy==1.24.3
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
httpx==0.25.2
loguru==0.7.2
pydantic-settings==2.1.0
pinecone-client
motor==3.3.2
pymongo==4.6.0